third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_delete_index5-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.delete() - index - throw InvalidStateError when the cursor is being iterated 
Harness: the test ran to completion.
